+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Change Log
 
 * DataObject `id` attribute is now a `uuid.UUID` object.
 * The second argument to a property directive may be a substitution reference
+* Added a code example for narrative resource discovery.
 
 0.12.0
 ======